Se arregla la funcion de graficar en base al idExpediente de la sesion y se arreglan errores en contronCrecimiento
