+++ public_html/Control/control.php
@@ -25,15 +25,15 @@
     </div>
     <script>
         const navItems = [
-            {name: "Vacunas", uri: "../Vacunas/vacunas.html"},
-            {name: "Estudios", uri: "../Estudios/estudios.html"},
-            {name: "Alergias", uri: "../Alergias/alergias.html"},
-            {name: "Antecedentes", uri: "../Antecedentes/antecedentes.html"},
-            {name: "Signos Vitales", uri: "../Signos/Signos.html"},
-            {name: "Tratamiento", uri: "../Tratamiento/Tratamiento.html"},
-            {name: "Procedimiento", uri: "../Procedimiento/procedimiento.html"},
-            {name: "Control de Crecimiento", uri: "../Control/control.html"},
-            {name: "Gráficas", uri: "../Graficar/index.php"}
+            {name: "Vacunas", uri: "..//Vacunas/vacunas.html"},
+            {name: "Estudios", uri: "..//Estudios/estudios.html"},
+            {name: "Alergias", uri: "..//Alergias/alergias.html"},
+            {name: "Antecedentes", uri: "..//Antecedentes/antecedentes.html"},
+            {name: "Signos Vitales", uri: "..//Signos/Signos.html"},
+            {name: "Tratamiento", uri: "..//Tratamiento/Tratamiento.html"},
+            {name: "Procedimiento", uri: "..//Procedimiento/procedimiento.html"},
+            {name: "Control de Crecimiento", uri: "..//Control/control.html"},
+            {name: "Gráficas", uri: "..//Graficar/index.php"}
 
         ];
     
@@ -44,7 +44,7 @@
         <h1 class="titulo-control mb-4">Control de crecimiento</h1>
         
         <div class="input-group mb-3">
-            <input type="text" class="form-control" placeholder="Buscar por Fecha" id="input-busqueda">
+            <input type="date" class="form-control" placeholder="Buscar por Fecha" id="input-busqueda">
             <button class="btn btn-outline-secondary" type="button" onclick="buscar()">Buscar</button>
             <button class="btn btn-outline-secondary" type="button" onclick="limpiar()">Limpiar</button>
         </div>
